+++ client/src/Pages/Chats/SoloPanel.js
