+++ client/src/Pages/Home/home.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Header from "../componets/Header/index";
-import Animations from "../componets/Animations/Animations";
-import GamesContainer from "../componets/GamesContainer/GamesContainer";
+import Header from "../../componets/Header/index";
+import Animations from "../../componets/Animations/Animations";
+import GamesContainer from "../../componets/GamesContainer/GamesContainer";
 // import "./home.css";
 
 const home = () => {
